Fix manual OAI XML import
At least in edu-sharing 7.x and 8.x, it is not possible to import an IEEE LOM file manually through the admin interface. When tried, Tomcat returns a generic `415 Unsupported Media Type` error. This is caused by the missing `@Consumes` statement for the given route, since a default of `application/json` is assumed.

Therefore, this commit fixes the erroneous behavior by allowing the correct content-type of `multipart/form-data`. This content-type is later expected in the code, such as by the call to `@FormDataParam("xml")` (returning the XML submitted).